# Improve performance of `first_value` by implementing special `GroupsAccumulator`