@2010YOUY01 2025-03-25
I haven't been following the recent conversations regarding hashmap optimization, but I also feel, if it needs pre-aggregate to make the low-cardinality case run faster, there might be some inefficiency inside the global state update implementation.
Since this PR already got significant performance improvement, we can play around with the further optimization in follow-up PRs, I plan to merge it shortly unless there are other considerations.